requirements.txt
# Web 框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 异步HTTP客户端
aiohttp==3.9.1
httpx==0.25.2

# 数据库
sqlalchemy==2.0.23
alembic==1.13.1

# 任务调度
apscheduler==3.10.4

# 日志
structlog==23.2.0
colorama==0.4.6

# 配置管理
python-dotenv==1.0.0
pyyaml==6.0.1

# 加密
cryptography>=41.0.0

# 浏览器自动化
selenium==4.15.2
playwright==1.40.0

# 工具库
click==8.1.7
rich==13.7.0
python-multipart==0.0.6

# 通知
email-validator==2.1.0

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# 代码质量
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0

# 开发工具
pre-commit==3.6.0
watchdog==3.0.0

# 数据处理
pandas==2.1.4
openpyxl==3.1.2

# 时间处理
python-dateutil==2.8.2
pytz==2023.3

# 网络请求
requests==2.31.0
urllib3==2.1.0

# JSON处理
orjson==3.9.10

# 环境变量
python-decouple==3.8

# 系统信息
psutil==5.9.6 